+++ src/client/java/babekon/sun/BabekonsSunPanelsClient.java
@@ -5,6 +5,6 @@ import net.fabricmc.api.ClientModInitializer;
 public class BabekonsSunPanelsClient implements ClientModInitializer {
 	@Override
 	public void onInitializeClient() {
-		// No GUIs registered
+
 	}
 }