+++ src/client/java/babekon/sun/screen/BatteryScreen.java
@@ -1,4 +1,0 @@
-package babekon.sun.screen;
-
-// GUI removed: no client screens.
-public class BatteryScreen {}