+++ src/client/java/babekon/sun/screen/SolarPanelScreen.java
@@ -1,4 +1,3 @@
 package babekon.sun.screen;
 
-// GUI removed: no client screens.
 public class SolarPanelScreen {}